rgpdd: an R interface to the Global Population Dynamics Database

The GPDD was initially compiled by John Prendargast, Ellen Bazeley-White?, Owen Smith, John Lawton and Pablo Inchausti and released in 1999. Version 2.0 was released in 2010 following a substantial restructuring of the database and the addition of 123 new series by David Kidd and Sarah Knight.

Details

Understanding the way in which populations of wild plants and animals behave over long periods of time is crucial to unravelling the way in which communities are assembled and the way in which they respond to disturbance, control or harvesting. The implications for conservation and agriculture are legion. Aside from practicalities, population variation is also intrinsically interesting, and provides a wealth of opportunity for mathematical innovation or exploration, especially when populations have particular cyclic, outbreaking or chaotic properties. For most students of population behaviour, the limiting factor in investigating any of these phenomena, and the development of theory to explain them, is the availability of suitable data. Usually, where analyses are performed and published, authors work on data that they have collected themselves. By definition, the collection of population time series is a lengthy process, and many ecologists have committed themselves to a lifetime of work in order to accumulate detailed information on populations at certain sites over many years.

Studies of population behaviour tend to address a number of themes, each with a typical taxonomic flavour. Thus, students of the chaotic vs cyclic question tend to focus on small mammals, those with an interest in the effects of harvesting or culling generally study fisheries or large mammal data respectively and analyses of insect populations tend to dominate the literature on pest control.

Analysis and subsequent publication can only take place once time series of adequate length have been amassed, but frequently authors will continue to collect data after publication and may follow the first paper with an updated or extended version, or a book or book chapter at a later date. There are examples of data sets that have been assiduously collected but from which no publications have resulted, or from which internal, private or unpublished documents have been generated.

The result of all this fragmentary activity in population dynamics, where data sets are often analysed individually, or in line with certain taxonomic conventions, is that it has been difficult to a) formulate general theory and b) investigate large scale pattern, both spatially and taxonomically. The general unavailability of data has also led to the development of theory through the repeated analysis of the same data sets. The celebrated Canadian lynx/snowshoe hare cycle has been the subject of analyses and publications almost too numerous to count. There is an obvious danger that if individual data sets such as this happen not to be representative of the way in which most populations behave then theoretical understanding may suffer.